producer/requirements.txt
kafka-python
google-cloud-storage
pandas
python-dotenv
